Separated key registration.
